requirements.txt
alembic==1.0.11
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.14
MarkupSafe==1.1.1
psycopg2==2.8.3
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
